+++ foxes/output/grids.py
@@ -54,7 +54,8 @@ def calc_point_results(
 
 def get_grid_xy(
     farm_results,
-    resolution,
+    resolution=None,
+    n_img_points=None,
     xmin=None,
     ymin=None,
     xmax=None,
@@ -62,6 +63,8 @@ def get_grid_xy(
     z=None,
     xspace=500.0,
     yspace=500.0,
+    states_sel=None,
+    states_isel=None,
     verbosity=0,
 ):
     """
@@ -74,6 +77,8 @@ def get_grid_xy(
         dimensions (state, turbine)
     resolution: float
         The resolution in m
+    n_img_points: tuple of int, optional
+        The number of image points (n, m) in the two directions
     xmin: float
         The min x coordinate, or None for automatic
     ymin: float
@@ -88,6 +93,10 @@ def get_grid_xy(
         The extra space in x direction, before and after wind farm
     yspace: float
         The extra space in y direction, before and after wind farm
+    states_sel: list, optional
+        Reduce to selected states
+    states_isel: list, optional
+        Reduce to the selected states indices
     verbosity: int, optional
         The verbosity level
 
@@ -103,8 +112,11 @@ def get_grid_xy(
         The grid points, shape: (n_states, n_pts, 3)
 
     """
-
     # prepare:
+    if states_isel is not None:
+        farm_results = farm_results.isel({FC.STATE: states_isel})
+    if states_sel is not None:
+        farm_results = farm_results.sel({FC.STATE: states_sel})
     n_states = farm_results[FV.H].shape[0]
 
     # get base rectangle:
@@ -114,11 +126,21 @@ def get_grid_xy(
     x_max = xmax if xmax is not None else farm_results[FV.X].max().to_numpy() + xspace
     y_max = ymax if ymax is not None else farm_results[FV.Y].max().to_numpy() + yspace
     z_max = z if z is not None else farm_results[FV.H].max().to_numpy()
+    
+    # compute number of points:
+    if resolution is not None and n_img_points is None:
+        nx = int((x_max - x_min) / resolution + 0.5) + 1
+        ny = int((y_max - y_min) / resolution + 0.5) + 1
+    elif resolution is None and n_img_points is not None:
+        nx, ny = n_img_points
+    else:
+        raise ValueError(f"Expecting either 'resolution' or 'n_img_points', got: resolution = {resolution}, n_img_points = {n_img_points}")
 
+    # compute points:
     x_pos, x_res = np.linspace(
         x_min,
         x_max,
-        num=int((x_max - x_min) / resolution) + 1,
+        num=nx,
         endpoint=True,
         retstep=True,
         dtype=None,
@@ -126,7 +148,7 @@ def get_grid_xy(
     y_pos, y_res = np.linspace(
         y_min,
         y_max,
-        num=int((y_max - y_min) / resolution) + 1,
+        num=ny,
         endpoint=True,
         retstep=True,
         dtype=None,
@@ -158,7 +180,8 @@ def get_grid_xy(
 
 def get_grid_xz(
     farm_results,
-    resolution,
+    resolution=None,
+    n_img_points=None,
     x_direction=270,
     xmin=None,
     zmin=0.0,
@@ -167,6 +190,8 @@ def get_grid_xz(
     y=None,
     xspace=500.0,
     zspace=500.0,
+    states_sel=None,
+    states_isel=None,
     verbosity=0,
 ):
     """
@@ -177,8 +202,10 @@ def get_grid_xz(
     farm_results: xarray.Dataset
         The farm results. The calculated variables have
         dimensions (state, turbine)
-    resolution: float
+    resolution: float, optional
         The resolution in m
+    n_img_points: tuple of int, optional
+        The number of image points (n, m) in the two directions
     x_direction: float
         The direction of the x axis, 0 = north
     xmin: float
@@ -195,6 +222,10 @@ def get_grid_xz(
         The extra space in x direction, before and after wind farm
     zspace: float
         The extra space in z direction, below and above wind farm
+    states_sel: list, optional
+        Reduce to selected states
+    states_isel: list, optional
+        Reduce to the selected states indices
     verbosity: int, optional
         The verbosity level
 
@@ -212,6 +243,10 @@ def get_grid_xz(
     """
 
     # prepare:
+    if states_isel is not None:
+        farm_results = farm_results.isel({FC.STATE: states_isel})
+    if states_sel is not None:
+        farm_results = farm_results.sel({FC.STATE: states_sel})
     n_states, n_turbines = farm_results[FV.H].shape
     n_x = np.append(wd2uv(x_direction), [0.0], axis=0)
     n_z = np.array([0.0, 0.0, 1.0])
@@ -236,10 +271,20 @@ def get_grid_xz(
     y_max = y if y is not None else np.max(yy)
     del xx, yy, zz
 
+    # compute number of points:
+    if resolution is not None and n_img_points is None:
+        nx = int((x_max - x_min) / resolution + 0.5) + 1
+        nz = int((z_max - z_min) / resolution + 0.5) + 1
+    elif resolution is None and n_img_points is not None:
+        nx, nz = n_img_points
+    else:
+        raise ValueError(f"Expecting either 'resolution' or 'n_img_points', got: resolution = {resolution}, n_img_points = {n_img_points}")
+
+    # compute points:
     x_pos, x_res = np.linspace(
         x_min,
         x_max,
-        num=int((x_max - x_min) / resolution) + 1,
+        num=nx,
         endpoint=True,
         retstep=True,
         dtype=None,
@@ -247,7 +292,7 @@ def get_grid_xz(
     z_pos, z_res = np.linspace(
         z_min,
         z_max,
-        num=int((z_max - z_min) / resolution) + 1,
+        num=nz,
         endpoint=True,
         retstep=True,
         dtype=None,
@@ -279,7 +324,8 @@ def get_grid_xz(
 
 def get_grid_yz(
     farm_results,
-    resolution,
+    resolution=None,
+    n_img_points=None,
     x_direction=270,
     ymin=None,
     zmin=0.0,
@@ -288,6 +334,8 @@ def get_grid_yz(
     x=None,
     yspace=500.0,
     zspace=500.0,
+    states_sel=None,
+    states_isel=None,
     verbosity=0,
 ):
     """
@@ -298,8 +346,10 @@ def get_grid_yz(
     farm_results: xarray.Dataset
         The farm results. The calculated variables have
         dimensions (state, turbine)
-    resolution: float
+    resolution: float, optional
         The resolution in m
+    n_img_points: tuple of int, optional
+        The number of image points (n, m) in the two directions
     x_direction: float
         The direction of the x axis, 0 = north
     ymin: float
@@ -316,6 +366,10 @@ def get_grid_yz(
         The extra space in y direction, before and after wind farm
     zspace: float
         The extra space in z direction, below and above wind farm
+    states_sel: list, optional
+        Reduce to selected states
+    states_isel: list, optional
+        Reduce to the selected states indices
     verbosity: int, optional
         The verbosity level
 
@@ -333,6 +387,10 @@ def get_grid_yz(
     """
 
     # prepare:
+    if states_isel is not None:
+        farm_results = farm_results.isel({FC.STATE: states_isel})
+    if states_sel is not None:
+        farm_results = farm_results.sel({FC.STATE: states_sel})
     n_states, n_turbines = farm_results[FV.H].shape
     n_x = np.append(wd2uv(x_direction), [0.0], axis=0)
     n_z = np.array([0.0, 0.0, 1.0])
@@ -357,10 +415,20 @@ def get_grid_yz(
     x_max = x if x is not None else np.max(xx)
     del xx, yy, zz
 
+    # compute number of points:
+    if resolution is not None and n_img_points is None:
+        ny = int((y_max - y_min) / resolution + 0.5) + 1
+        nz = int((z_max - z_min) / resolution + 0.5) + 1
+    elif resolution is None and n_img_points is not None:
+        ny, nz = n_img_points
+    else:
+        raise ValueError(f"Expecting either 'resolution' or 'n_img_points', got: resolution = {resolution}, n_img_points = {n_img_points}")
+
+    # compute points:
     y_pos, y_res = np.linspace(
         y_min,
         y_max,
-        num=int((y_max - y_min) / resolution) + 1,
+        num=ny,
         endpoint=True,
         retstep=True,
         dtype=None,
@@ -368,7 +436,7 @@ def get_grid_yz(
     z_pos, z_res = np.linspace(
         z_min,
         z_max,
-        num=int((z_max - z_min) / resolution) + 1,
+        num=nz,
         endpoint=True,
         retstep=True,
         dtype=None,