+++ foxes/utils/xarray_utils.py
@@ -37,12 +37,12 @@ def write_nc(ds, fpath, round={}, complevel=9, verbosity=1, **kwargs):
                     print(f"File {fpath.name}: Rounding {v} to {d} decimals")
                 ds[v].data = ds[v].data.round(decimals=d)
 
-    if verbosity > 0:
-        print("Writing file", fpath)
-
     enc = None
     if complevel is not None and complevel > 0:
         if verbosity > 1:
             print(f"File {fpath.name}: Compression level = {complevel}")
         enc = {k: {"zlib": True, "complevel": complevel} for k in ds.data_vars}
+
+    if verbosity > 0:
+        print("Writing file", fpath)
     ds.to_netcdf(fpath, encoding=enc, **kwargs)